src/utils/media_processor.py
"""
媒体处理器 - 处理各种媒体文件的提取和转换
"""

from typing import Dict, Any, Optional
from pathlib import Path
import logging

logger = logging.getLogger(__name__)


class MediaProcessor:
    """媒体处理器类，用于处理视频、音频、文档等文件"""
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        self.config = config or {}
    
    def process_file(self, file_path: str, media_type: str) -> Dict[str, Any]:
        """
        处理媒体文件
        
        Args:
            file_path: 文件路径
            media_type: 媒体类型 (video, audio, document, text)
            
        Returns:
            处理结果字典
        """
        logger.info(f"Processing {media_type} file: {file_path}")
        
        if media_type == "video":
            return self._process_video(file_path)
        elif media_type == "audio":
            return self._process_audio(file_path)
        elif media_type == "document":
            return self._process_document(file_path)
        elif media_type == "text":
            return self._process_text(file_path)
        else:
            raise ValueError(f"Unsupported media type: {media_type}")
    
    def _process_video(self, file_path: str) -> Dict[str, Any]:
        """处理视频文件"""
        # 这里应该实现视频处理逻辑
        # 包括提取关键帧、转录音频等
        logger.info(f"Processing video file: {file_path}")
        return {
            "file_path": file_path,
            "media_type": "video",
            "status": "processed",
            "frames": [],
            "transcript": "",
            "duration": 0.0
        }
    
    def _process_audio(self, file_path: str) -> Dict[str, Any]:
        """处理音频文件"""
        # 这里应该实现音频处理逻辑
        # 包括语音识别等
        logger.info(f"Processing audio file: {file_path}")
        return {
            "file_path": file_path,
            "media_type": "audio",
            "status": "processed",
            "transcript": "",
            "duration": 0.0
        }
    
    def _process_document(self, file_path: str) -> Dict[str, Any]:
        """处理文档文件"""
        # 这里应该实现文档处理逻辑
        # 包括PDF、DOCX等格式的解析
        logger.info(f"Processing document file: {file_path}")
        try:
            # 尝试从文档中提取内容
            import os
            file_ext = os.path.splitext(file_path)[1].lower()
            
            if file_ext == ".pdf":
                # 尝试使用PyMuPDF提取文本
                try:
                    import fitz
                    doc = fitz.open(file_path)
                    content = ""
                    for page in doc:
                        content += page.get_text()
                    doc.close()
                    return {
                        "file_path": file_path,
                        "media_type": "document",
                        "status": "processed",
                        "content": content,
                        "pages": len(doc)
                    }
                except Exception as e:
                    logger.warning(f"Failed to extract PDF content: {str(e)}")
                    return {
                        "file_path": file_path,
                        "media_type": "document",
                        "status": "processed",
                        "content": f"PDF文件: {os.path.basename(file_path)}",
                        "pages": 0
                    }
            elif file_ext == ".txt":
                # 处理文本文件
                try:
                    with open(file_path, 'r', encoding='utf-8') as f:
                        content = f.read()
                    return {
                        "file_path": file_path,
                        "media_type": "document",
                        "status": "processed",
                        "content": content,
                        "pages": 1
                    }
                except Exception as e:
                    logger.warning(f"Failed to read text file: {str(e)}")
                    return {
                        "file_path": file_path,
                        "media_type": "document",
                        "status": "processed",
                        "content": f"文本文件: {os.path.basename(file_path)}",
                        "pages": 1
                    }
            elif file_ext == ".docx":
                # 处理Word文档
                try:
                    from docx import Document
                    doc = Document(file_path)
                    content = ""
                    for para in doc.paragraphs:
                        content += para.text + "\n"
                    return {
                        "file_path": file_path,
                        "media_type": "document",
                        "status": "processed",
                        "content": content,
                        "pages": len(doc.paragraphs)
                    }
                except Exception as e:
                    logger.warning(f"Failed to extract DOCX content: {str(e)}")
                    return {
                        "file_path": file_path,
                        "media_type": "document",
                        "status": "processed",
                        "content": f"Word文档: {os.path.basename(file_path)}",
                        "pages": 0
                    }
            else:
                # 其他文档类型
                return {
                    "file_path": file_path,
                    "media_type": "document",
                    "status": "processed",
                    "content": f"文档文件: {os.path.basename(file_path)}",
                    "pages": 0
                }
        except Exception as e:
            logger.error(f"Error processing document: {str(e)}")
            return {
                "file_path": file_path,
                "media_type": "document",
                "status": "error",
                "content": f"文档处理失败: {str(e)}",
                "pages": 0
            }
    
    def _process_text(self, file_path: str) -> Dict[str, Any]:
        """处理文本文件"""
        try:
            with open(file_path, 'r', encoding='utf-8') as f:
                content = f.read()
            return {
                "file_path": file_path,
                "media_type": "text",
                "status": "processed",
                "content": content,
                "char_count": len(content)
            }
        except Exception as e:
            logger.error(f"Error processing text file: {e}")
            return {
                "file_path": file_path,
                "media_type": "text",
                "status": "error",
                "error": str(e)
            }
    
    def extract_text(self, file_path: str, media_type: str) -> str:
        """
        从媒体文件中提取文本
        
        Args:
            file_path: 文件路径
            media_type: 媒体类型
            
        Returns:
            提取的文本内容
        """
        result = self.process_file(file_path, media_type)
        
        if result.get("status") == "error":
            return ""
        
        if media_type in ["text", "document"]:
            return result.get("content", "")
        elif media_type in ["video", "audio"]:
            return result.get("transcript", "")
        
        return ""
