requirements.txt
# Core Dependencies
openai>=1.12.0
crewai>=0.28.0
crewai-tools>=0.0.15
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Video/Audio Processing
opencv-python>=4.8.0
moviepy>=1.0.3
pydub>=0.25.1
speechrecognition>=3.10.0
pyaudio>=0.2.13
openai-whisper>=20230314
librosa>=0.10.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.13.0

# API
fastapi>=0.104.1
uvicorn>=0.23.2
python-multipart>=0.0.6

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
dash>=2.14.0
streamlit>=1.29.0

# Utilities
numpy<2.0.0
requests>=2.31.0
aiohttp>=3.9.0
tenacity>=8.2.0
tqdm>=4.66.0
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0